chore: update paperclip and lock async_trait
Point paperclip to our own repo with the latest fixes until a new
release is published to crates-io.
Lock async_trait because of an imcompatibility with the tracing crate.
Make the mayastor derivation depend on the $rev so that nix is aware of
this and knows when it needs to rebuild it.
Set panics to abort on dev builds.