feat(store keys): define keys for store resources
Define the keys that should be used for resources that can be saved in
the persistent store.

The keys are defined for the 'spec' and 'state' types for each resource.

Additional changes:
Reorganise the directory structure of the store types. Each type is
moved out of v0.rs and placed in its own source file.