feat: use transactions for modifying replicas
Adds a transaction record for updating resources. This aims to give the
core agent more information to make better decisions when things go wrong.

Retry writing to the persistent store in a separate worker thread and return
an error back to the caller if other modifying operation is tried.

Add tests to encompass this transaction model which sadly have to rely
on timing for the moment.
todo: investigate mocking libraries for rust - ideally we'd want to fake
the whole core agent to test all possible failures.