refactor: update the volume create api
This follows the new volume api described in the design docs.
Topology information may now be filled in via the operator, either as an explicit
selection of nodes, or using labels for nodes and pools.

Before the labels can be used the node and pool will have to be updated to include
them at creation time and also via an update method.