feat(test): add docker-compose
This adds an example docker-compose file which can be used to spin up
the control plane components as well as mayastor instance. This file is
intended to serve as an example to and is similar to the deployer tool.
The deployer tool uses the composer to pro grammatically construct a
cluster where is this is more static.