feat(volume): add new volume operations
Add share/unshare and publish/unpublish volume operations.
The publish allows for a nexus to be created or a specified node or any suitable node. In
addition, a share protocol may also be specified.

Volume creation does not result in a nexus being created now. A nexus shall only exist
if the volume is published.

The spec update operations are now generic enough that the code can be shared making use of
the existing traits.
todo: spec creation should also be made generic!
todo: operation validation should also be made generic!