feat: bootstrap the registry resource specs
Make the registry initialise the resource specs with the contents of the
persistent store. This ensures that the specs in the registry match what
was there prior to the core agent being restarted. This makes the core
agent more resilient across restarts.

The resource specs are also now exposed through the REST API. This
allows the caller to inspect the requested configuration of the various
resources. If any resource is unhealthy, the spec will be the
configuration that the reconciler will attempt to achieve.

Additional changes:
- Move type definitions to their own directory to avoid cyclical
dependencies.