feat: add nexus child abstraction for the nexus spec
The nexus spec was previously following the nexus status, where all children are simply
ChildUris/Strings.
To better accommodate the control plane logic the nexus children are now
either a Replica or a RawUri, allowing us to make better decisions as we'll know exactly
which replica is part of a nexus without having to parse the uri.