feat: add SetVolumeReplica request handler to the core agent
Add new increase/decrease methods, which allows us to increase/decrease the replica
count of a volume (depending on the setvolumereplica request).

To achieve this, we make use of the new replica scheduling functionality added previously.
The creation of a volume nexus is now also making use of the new scheduling logic and the
{Add/Remove}NexusReplica requests.

For simplicity, only a change of 1 replica at a time is allowed. This limitation will be
removed we have a fully fledged reconciler.