fix: create nexuses with only healthy replicas
When mayastor nexus faults a child it updates its persistent nexus information on etcd to
mark the child as not healthy.
Also when the nexus is not shutdown cleanly, then the clean flag remains unset on etcd.

The control plane now inspects this information on etcd before attempting to create a nexus.
It selects a list of replicas for nexus creation and filters them down according to the
information it retrieved from etcd.

Added tests for various cases where replicas are faulted. These faults are mocked.

Change node_wrapper getter to return an error if the node cannot be found.

Fix tests to use uuid as a query parameter when creating a nexus.