feat: add node spec to the persistent store
Write out the node spec to the persistent store.
On startup the core agent can use this to attempt to load all node information if the
nodes are online.

Changed the openapi spec to return the Node with a spec and state.
Rename the state/status for the nodes properly.