fix: don't remove too many replicas
When removing a replica from a volume or a child from a nexus we could
potentially "too many" and cause the volume to become degraded, or worst, faulted.
Now, when we remove a replica we always check that we're not falling into that case.

Note: we can still get things wrong because the state of nexuses and its children
is volatile so our decisions could be based on stale information.
To address this better we need cas -1018.