fix: fixup the release script
Build the core and jsongrpc agents independently so that the release
script is able to load the individual tar files. Building them together
resulted in only the first (core) agent being loaded.

Note: REGISTRY_IP should be replaced with the desired Docker registry IP
before attempting to run the script.