feat(CI): build and push control plane images
Enable CI to build and push control plane images for certain branches.

Build core and jsongrpc images separately to allow both to be loaded
from the resulting tar file. Attempting to build them both together was
resulting in only the core image being loaded from the tar file.

Provide deployment YAML files for the core agent and REST server. The
REST server is exposed through a NodePort service on port 30010.

Bug fix:

- Allow the etcd enpoint to be provided to the core agent without
the need for a port to be specified. Omitting the port results in the
default port being chosen (2379). This mimicks the behaviour of
Mayastor.