feat(kubectl plugin): plugin to make REST calls
Create a kubectl plugin which can interact with the new control plane
through REST calls.

The REST calls are made through the use of the in-tree OpenAPI client
which is autogenerated from the OpenAPI spec.

Currently the plugin only works over HTTP (not HTTPS) as the appropriate
certificates haven't been set up. Because of this, the REST deployment
and REST service have had their YAML files updated to expose the HTTP
port on NodePort 30011.

The plugin currently supports the following operations:

- Getting all volumes
- Getting a specific volume by ID
- Scaling a specific volume (increasing/decreasing replica count)
- Getting all pools
- Getting a specific pool by ID