chore: remove "ANA bits" from the current volume
A non-ANA volume is only meant to have one active nexus target at a time.
ANA adds a lot more complexity than than just a vector of nexuses, so it's better
to completely ignore multi volume nexuses for now and add it back later as a new
type of volume.
This way we can also make sure that we don't break non-ANA volumes when we start
adding ANA volumes..