feat: consider a node online even if the watchdog expired
If there is some kind of temporary nats issue the heartbeat will fail and the node
will be set as offline.
This means that offline may not always mean offline and we can't use this state
to avoid issuing client requests.

Now, when the HB expires we issue a liveness probe to assess is the node is online
or not, and if it is then rearm the watchdog and keep the node online.