chore: use the "latest" mayastor
Mayastor now uses the proto api as a submodule and this makes it hard to use
it as a derivation as the sha256 seems to change...

Instead, use a docker image by default. A specific binary mnay also be
specified on the nix-shell via the mayastor argument or you can also
specify either image or binary via the deployer.