chore: update helm to add init containers
Adds initcontainers for nats and etcd to all control plane images for now.
This mitigates some of the issues where one component starts to work before
the core agent is ready, which in turns depends on etcd and nats.

todo: Ideally we should instead add liveness and readyness probes instead.