chore(msp): add msp crd yaml file to the deploy
Update the msp with a flag to generates the crd yaml file and exit.
Update the msp to include pool capacity and free space.
Add script that generates the crd file when the msp src code changes.
Run the crd generation on pre-commit.
Run the helm-deploy generation on pre-commit.
Modify the msp poll interval to match the common cache update interval.
Add TERM trap to the init containers.