fix: protocol should not be part of the volume
The share protocol should not be part of the volume as it's decided on publish.
In the future we may even do things such as publish the volume with a protocol
that is suited for the node where the application runs.

So, instead make the protocol part of the volume target (ie the nexus).

Also, make the state a non optional component of the volume so that we do not
need to "fake" it on multiple layers, instead the core agent will set it.