feat(Volume info): add replica topology to volumes
Add replica topology information when returning volumes through the REST
API.

The topological information is represented as a hash map where the key
is the replica UUID and the value is a structure representing the
toplogy. The topology includes the node and pool where the replica
resides and the state of the replica.

The kubectl plugin now supports an additional "replica-topology" command
which returns the replica topology for a given volume.