test: add CARGO_PKG_NAME and test module to trace spans
Add env CARGO_PKG_NAME to the traces and also the test module name.
The test module name is figured out from a backtrace where we go down
until we see a source code from our repo that is not the test lib. This
is a bit ugly but doesn't look like there is any other way of getting it.