fix: make sure multiple states are viewed consistently
When we fetch multiple specs and state objects, things might change "in between"
as a result of current operations leading to potentially confusing results.

Add RW access to the states to make sure we get a consistent picture across
the cluster and its various resources.

When we start to get the volume spec don't allow anyone else to attempt modify it
If an update is in progress, return the stale but consistent value of the spec