fix(pool CRD): set to unknown when pool is missing
A pool CRD should appropriately reflect the state of a pool if the node
on which it resides becomes inaccessible.

When a node becomes inaccessible the control plane is unable to retrieve
any runtime state information for resources on that node. As a result,
the cached state information is invalidated (set to None). If a REST
request is made for a resource on an inaccessible node, the returned
object will not contain any state information.

If the MSP operator receives a pool object (via REST) without a state,
it cannot determine the health of the pool. Consequently, it will now
reflect this in the CRD as an 'Unknown' state.

The MSP operator will periodically check if the pool state is available.
Once this is received, the state of the CRD will be updated
appropriately (that is, it will be set to Online if the pool returns in
a healthy state).