feat(failover): allow the core-agent to failover
Use an etcd lease lock to allow the core-agent to failover when its node
has crashed/disappeared/*.

We now allow core-agent failover by using lease locks to check if a previous
instance of the core-agent was still running. This *NOT* a distributed lock,
but simply a means of failover from an instance of `instance_name` to another
(eg: when using a kubernetes deployment of *1* replica).

When another instance of `instance_name` appears and take over the lock we shall
*panic* as in this case we're not intending to keep the previous owner running.
Only one instance of `instance_name` should run at a time!

If a previous instance of the core-agent somehow comes back up and attempts to
write to etcd it will fail as its lost its lease.