fix(MSP operator): remove assertion
The MSP operator was making the assumption that an older version of a
resource would always have a lower resource version number than a newer
version of the resource. This turns out not to be the case. From the
documentation
(https://kubernetes.io/docs/reference/using-api/api-concepts/#resource-versions):

"...clients must not assume resource versions are numeric, and may only
compare two resource versions for equality (i.e. must not compare
resource versions for greater-than or less-than relationships)."