refactor(images): tidy up building of images
There's a new build type at the root of the images which should simplify
adding new images (no need for -dev or -cov). Example, to build a debug
version of the core agent:
nix-build -A images.debug.agents.core

Add a build suffix to fix the debug builds which are broken. Example:
the core agent image will be named mcp-core and the debug mcp-core-dev