refactor(common lib): create new utils crate
The constant definitions and building of the constants YAML file have
been moved from the common-lib crate to a new utils crate. In addition
to this the utils crate provides function and macro utilities that can
be used by other crates.

The purpose of splitting out the functionality is to improve build
times of other crates if all they require are the constant definitions.
Using the common-lib crate required pulling in a number of unnecessary
dependencies. Now these crates can instead use the utils crate which is
much lighter-weight (has fewer dependencies).