fix(rust test): fix occasional CI test failures
CI occasionally shows the "hotspare" test failing; specifically in the
"hotspare_replica_count_spread" function.

The problem appears to be in the "wait_till_volume" function where we
were waiting for the number of replicas to be correct. However, we were
checking the replica count in the spec, which defines the desired number
of replicas rather than the actual number of replicas in the system.

This function has been enhanced to check the replica topology field in
the volume state which will give the number of actual replicas.