fix: sending node online event should not block
On startup if the cluster had more than 5 Nodes the send event notification per node would block.
This is because the queue has only 5 elements.
Simply increasing the queue size would not necessarily improve things so instead we drop the
notification if the queue is full which is acceptable as at the moment we have no way of
"flattening" the events or to avoid running the reconcilers too often so we probably don't want
a huge queue of notifications to be built. Losing one is not a problem, just may mean we'd be a
little slower.