chore: don't trace reconciliation unless TRACE is set or there is work to do
For clusters with high number of volumes there are a lot of traces being sent that sometimes
seem to overwhelm the telemetry exporter.
During reconciliation we should not create spans needlessly unless we're on TRACE level or unless
the reconciler actually did work.
TODO: We need a generic strategy to cater for this