fix(nexus): recreate faulted nexuses when appropriate
When the healthy replicas of a faulted nexus are Online we should attempt to recreate said nexuses.
This is done by first destroying the faulted nexus. The missing nexus reconciler will then recreate
the nexus.

When no healthy children are available, the faulted nexus are left behind until such time the node
crashes or the volume is unpublished.

Resolves: CAS-1296