fix(deployer): deploying mayastor binary is fixed
Mayastor now links against `libspdk-bundle.so`, which is located in the same
directory as binary itself (previously, it linked against libspdk.so from
a Nix package). Finding this shared library requires access to the directory
with mayastor binary from within the container. Support for that has been
added to the `composer` crate.
This fix does not affect deploying mayastor from image.

Fixes CAS-1290