fix: spawn core pool create/destroy to avoid early drop on timeout
The existing code does not handle being dropped very well, as it will leave things in a
potentially "bad" state.
TODO: handle serive handlers drops gracefully. This will be tricky as eg ETCD client is async and so
we can't really flush things to ETCD in a nice way on drop.

For now, try to avoid dropping the futures when the client times out by spawning it.
NOTE: the future can still be dropped on reactor shutdown.