feat(nexus info): delete from persistent store
Ensure the NexusInfo structure is deleted from the persistent store when
a volume is destroyed. Also delete old NexusInfo structures when a
volume is re-published.

Also provide the ability for the control plane to pass the key that
should be use to persist the NexusInfo structure to the data plane. The
key itself has been modified to include the volume and nexus UUID.