feat(pagination): volumes
Provide support for pagination when getting volumes. This should be used
when there are a large number of volumes which can cause a "get volumes"
request to time out.

The ResourceMap has been changed from a HashMap to an IndexMap. This
preserves the order of inserts when interating over the map, allowing
paginated results to return the correct entries over multiple calls.