chore: merge terraform and rebuild fixes
feat: terraform libvirt/lxd cluster and deployment

Ports the terraform scripts from the dataplane repo and tweak it so that it's ready to use
out of the box (mostly).

fix(rebuild): correct system wide rebuilds

The number of rebuilds returned by the data plane nexuses represents the
total number of rebuilds on that node; not the number of rebuilds per
nexus.

Ensure that we don't accumulate the same rebuilds multiple times by
choosing a single nexus to retrieve this number from.