feat: drain work in progress
Signed-off-by: chriswldenyer <christopher.denyer@datacore.com>
Signed-off-by: chriswldenyer <christopher.denyer@mayadata.io>

Add APIs to drain a node.
Drain is not yet fully implemented - what is missing is the reconciler that moves nexuses to another node, which is dependent on HA being implemented.
With this PR, drain labels are now stored in the core agent and perform the same cordoning function as cordon labels but also flag that a drain is to be undertaken.
The Node object returned from the core-agent now includes a DrainStatus field which is one of:
NotDraining
Draining
Drained
dependent on the presence of nexuses and drain labels on the node.
The kubectl plugin initiates a drain via:
kubectl-plugin drain node
and cancels it via:
kubectl-plugin uncordon node

A new option --show-drain when used with "get node" or "get nodes" will display the drain labels and the state.
When used with "get nodes" only nodes that are in Draining or Drained state are listed.