feat: drain initial implementation
Signed-off-by: chriswldenyer <christopher.denyer@datacore.com>
Signed-off-by: chriswldenyer <christopher.denyer@mayadata.io>

Add APIs to drain a node.
Drain is not yet fully implemented - what is missing is the reconciler
that moves nexuses to another node, which is dependent on HA being
implemented.
With this PR, drain labels are now stored in the core agent and perform
the same cordoning function as cordon labels but also flag that a drain
is to be undertaken.
The Node object returned from the core-agent now includes a boolean
is_draining field.
This is set via the drain api command but will be cleared by the
reconciler, when implemented.

The kubectl plugin initiates a drain via:
    kubectl-plugin drain node
and cancels it via:
    kubectl-plugin uncordon node

Two new get objects cordon and drain are implemented for the plugin.

    get cordon <node id> # lists cordon information concerning the node
    get cordons          # lists cordon information for all cordoned nodes
    get drain <node id>  # lists drain information concerning the node
    get drains           # as above but for all nodes with drain labels

Note: the get node <node id> --show-cordon-labels option has been removed.