feat: drain initial implementation
Signed-off-by: chriswldenyer <christopher.denyer@mayadata.io>

Add APIs to drain a node.
Drain is not yet fully implemented - what is missing is the reconciler
that moves nexuses to another node, which is dependent on HA being
implemented.
With this PR, drain labels are now stored in the core agent and perform
the same cordoning function as cordon labels but also flag that a drain
is to be undertaken.
The Node object within the core-agent now includes a DrainState
enum variant which contains both cordon labels and drain labels.
This is controlled via the REST API but also by the reconciler,
when implemented.

The kubectl plugin initiates a drain via:
    kubectl-plugin drain node <node name> <drain label>
and cancels it via:
    kubectl-plugin uncordon node <node name> <drain label>

Two new get objects cordon and drain are implemented for the plugin.

    get cordon node <node id> # lists cordon information concerning the node
    get cordon nodes          # lists cordon information for all cordoned nodes
    get drain node <node id>  # lists drain information concerning the node
    get drain nodes           # as above but for all nodes with drain labels

Note: the "get node <node id> --show-cordon-labels" variant has been removed.