feat(cordon): implement cordon logic
Add cordon support to the Core Agent.

Applying a cordon prevents the control plane from being able to schedule
resources on that node.

When cordoning a node a label must be applied. This uniquely identifies
a particular cordon operation. When uncordoning a node the appropriate
label must also be supplied.

A particular cordon label can only be applied once. Attempting to apply
a cordon with a label that already exists will fail.

If the label supplied with the uncordon operation does not exist, an
error is returned.