feat(thin provisioning): add flag
Add a "thin" flag to the control plane which denotes whether or not a
volume should be thin provisioned.

Manually tested in k8s by provisioning a volume using a storage class
which has the "thin" parameter set to true.

The thin flag is included in the VolumeSpec and is persisted to etcd
along with all the other VolumeSpec parameters.

Note: This commit only supports the flag itself. Volumes are still only
created as thick provisioned.

Signed-off-by: Paul Yoong <paul.yoong@mayadata.io>