chore: add worker thread knobs to the api-rest service
By default set the max number of worker threads is set to 8.
This will unblock running on servers with large numbers of cpus and using more memory
than what our defaults in the helm charts are prepared to dish out.

Adds 2 new args to the api-rest, workers and max workers.
By default workers is the number of physical cpus and the default max is 8.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>