refactor: split SpecOps into separate traits
GuardedSpecOperations now is implemented only for the Guards.
A reduced SpecOperations is implemented only for the inner Specs.

todo: change all operations to receive the Guard as part of the call - this should
make the api more obvious and also make sure there are no modifications in case
someone forgets to setup the guard at the top level.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>