chore(cherry-pick): use rustup with path
On non-nixos when using rustup and the system's libc is not the same as the nix-shell's libc seems
 we get some odd errors:
https://discourse.nixos.org/t/nix-shell-with-rustup/22452
This is likely because the binaries from rustup haven't been patched to point to the nix env.

When this happens set rust-toolchain.toml to our cloned rust toolchain aso the IDE can work
correctly but use the RUSTUP_TOOLCHAIN under the nix-shell so we can use rustup properly.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>