fix(shutdown): don't fail when lazily shutting down the nexus
When the core agent restarts at the same time as the older nexus node, then we don't have
the node state and as such we fail the shutdown call.
For a republish call we want to shutdown in a lazy manner, so ignore this error.
The reconcilers will pick up the slack, if need be.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>