fix(switchover): destroy shutdown targets lazily
The cluster agent has some issues handling this itself, eg: it doesn't know when nodes are
online. It would have to ask the core-agent. Since it doesn't have this functionality yet
it's better to have the core agent handle these destroys lazily.

Also added a switchover test for this effect.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>