fix(ha): invalid nqn in mayastor-agent-ha-node logs
In current implementation NQNs for suspected NVMe paths are obtained
from a Subsystem instance constructed from a platform NVMe subsystem
which is in “connecting” state. Sometimes platform reports “(efault)”
as NQNs for such disconnected subsystems.
As a fix, NQN is now obtained at the moment when NVMe path appears in
the system for the first time.

Signed-off-by: Mikhail Tcymbaliuk <mtzaurus@gmail.com>