fix(agent-node): verify nqn when disconnecting
When attempting to disconnect a path, make sure that the nqn is the same.
Why? We found a bug where we would keep a stale path, and it's possible that a new device
may be reuse that path (new nqn ofc) and so we should ensure the nqn is the expected
when disconnecting to make sure we don't remove the wrong device!

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>