fix(agent-node/subsystem): handle multiple controller paths
The existing code seems to treat a subsystem as having a single path.
The nvmeadm library subsystem type seems to be quite confusing using the term controller
and subsystem interchangeably.
This poses a problem when swithching over as temporarily the subsystem will have two controllers,
the new one (good) and the old one (bad).

Here we can see the old controller nvme0 and the new one nvme1 by probing sysfs at the right time:
tiago@ksnode-2:~$ ls /sys/class/nvme-subsystem/nvme-subsys0/
model  ng0n1  nvme0n1 (block device is still 0)  nvme0 (old) nvme1 (new)  subsysnqn ..

What was then happening is that when fixing a path to live we were setting the suspected_path nvme0
as Good, rather than leave nvme0 as Bad and set nvme1 as Good.
This posed a problem on a subsequent failure as we would even try to disconnect nvme0 instead of
nvme1 - this could cause us to boot the wrong controller and causing issues on another disk!!!

This fix is to keep a list of paths per NQN on the detector. Additionally when we fetch a controller
per NQN we have to fetch a list instead (again, multiple paths for the same NQN...)

Ideally we should fix things properly (nvmeadm) but nearing the release this is probably less risky.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>