fix(pstor/etcd): include queuing as part of the timeout
As things stands, store operations are serialized. This is simply a side-effect of the etcd client
library taking operation methods in a mutable fashion.
todo: whilst we probably don't want 1000x requests to hit the store at once, we probably can open
the tap for more than 1 at a time :)

For now, simply include the queuing for our "slot" as part of the timeout.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>