fix: import pool when destroying
When a node is restarted, we may issue a pool destroy before the reconciler imports it.
Note: the reconciler will never attempt to import it after this as the pool will now be in
the Destroying phase.
To address this, if pool destroy fails with status NotFound, we then attempt to import it
and then retry the destruction.

Added a test case for this.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>