feat(capacity): remove largest enospc replicas
When a pool is exhausted multiple thin-provisioned replicas may fail with ENOSPC.
In this situation we'll currently adopt the simplistic strategy of faulting the largest replica
on the pool, or more specifically the replica with largest actual allocation.
A pre-condition for the replica faulting is that the volume to which the replica belongs to
should retain "enough" remaining healthy replicas! For example, we can't fault the last replica
of a volume as we can't rebuild from "thin" air.

todo: online nexus children if their respective pools have sufficient capacity.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>